requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1733753955715/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1759762205280/work
Automat @ file:///home/conda/feedstock_root/build_artifacts/automat_1744877909910/work
av==15.1.0
backoff==2.2.1
bce-python-sdk==0.9.52
bcrypt==5.0.0
beautifulsoup4==4.13.5
bitsandbytes==0.48.2
Brotli @ file:///croot/brotli-split_1736182456865/work
build==1.3.0
cachetools==6.2.2
camelot-py==1.0.9
certifi==2025.11.12
cffi @ file:///home/task_176009805548162/conda-bld/cffi_1760098067589/work
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
chromadb==1.3.5
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
constantly==15.1.0
contourpy==1.3.3
cryptography @ file:///home/task_176010605113983/conda-bld/cryptography-split_1760106309120/work
cssselect @ file:///home/conda/feedstock_root/build_artifacts/cssselect_1747322906730/work
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.1
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
docx2txt==0.9
durationpy==0.10
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
fastapi==0.115.6
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.59.2
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
ghostscript==0.8.1
google-api-core==2.25.1
google-auth==2.43.0
google-cloud-vision==3.10.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.74.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
hyperlink @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_hyperlink_1733319972/work
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental @ file:///home/conda/feedstock_root/build_artifacts/incremental_1734273526996/work
itemadapter @ file:///home/conda/feedstock_root/build_artifacts/itemadapter_1756838098870/work
itemloaders @ file:///home/conda/feedstock_root/build_artifacts/itemloaders_1734947448722/work
Jinja2==3.1.6
jmespath @ file:///home/conda/feedstock_root/build_artifacts/jmespath_1733229141657/work
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.29
langchain-core==0.3.75
langchain-huggingface==0.3.1
langchain-ollama==0.3.7
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.27
lxml @ file:///croot/lxml_1737039601731/work
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
modelscope==1.32.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
olefile==0.47
ollama==0.6.1
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.23.2
opencc-python-reimplemented==0.1.7
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt-einsum==3.3.0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.0
paddlex==3.3.9
panda==0.3.1
pandas==2.3.2
parsel @ file:///home/conda/feedstock_root/build_artifacts/parsel_1737145686686/work
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20250506
peft==0.17.1
pi_heif==1.1.0
pikepdf==9.11.0
pillow==11.3.0
posthog==5.4.0
prettytable==3.17.0
propcache==0.4.1
Protego @ file:///home/conda/feedstock_root/build_artifacts/protego_1750786835502/work
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyDispatcher==2.0.5
Pygments==2.19.2
PyMuPDF==1.26.4
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1760304524008/work
pyparsing==3.2.4
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
queuelib @ file:///home/conda/feedstock_root/build_artifacts/queuelib_1743447365992/work
qwen-vl-utils==0.0.11
RapidFuzz==3.14.1
readability-lxml==0.8.4.1
referencing==0.37.0
regex==2025.9.1
requests==2.32.5
requests-file @ file:///home/conda/feedstock_root/build_artifacts/requests-file_1760630329386/work
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
Scrapy @ file:///home/conda/feedstock_root/build_artifacts/scrapy_1751481171934/work
sentence-transformers==5.1.0
service-identity @ file:///home/conda/feedstock_root/build_artifacts/service-identity_1733366740127/work
setuptools==78.1.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.41.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.19
tldextract @ file:///home/conda/feedstock_root/build_artifacts/tldextract_1745323414969/work
tokenizers==0.22.1
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1760014679/work
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
trl==0.25.1
Twisted @ file:///croot/twisted_1743537782170/work
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
unstructured==0.18.14
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
w3lib @ file:///home/conda/feedstock_root/build_artifacts/w3lib_1737998860387/work
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.interface @ file:///io/buildout/croot/zope.interface_1735844903240/work
zstandard==0.25.0
